move data buffer to types
