+++ aether/methods/client_api/client_root_api.h
@@ -19,7 +19,7 @@
 
 #include "aether/types/uid.h"
 #include "aether/events/events.h"
-#include "aether/transport/data_buffer.h"
+#include "aether/types/data_buffer.h"
 #include "aether/api_protocol/api_class_impl.h"
 #include "aether/api_protocol/return_result_api.h"
