+++ aether/methods/server_reg_api/root_api.h
@@ -23,7 +23,7 @@
 #  include "aether/crypto/signed_key.h"
 #  include "aether/crypto/crypto_definitions.h"
 
-#  include "aether/transport/data_buffer.h"
+#  include "aether/types/data_buffer.h"
 #  include "aether/api_protocol/api_method.h"
 
 namespace ae {
