+++ aether/stream_api/safe_stream/safe_stream_api.h
@@ -19,7 +19,7 @@
 
 #include <cstdint>
 
-#include "aether/transport/data_buffer.h"
+#include "aether/types/data_buffer.h"
 #include "aether/api_protocol/api_method.h"
 #include "aether/api_protocol/api_class_impl.h"
 #include "aether/api_protocol/return_result_api.h"
