+++ aether/transport/low_level/tcp/data_packet_collector.h
@@ -26,7 +26,7 @@
 #include "third_party/aethernet-numeric/numeric/tiered_int.h"
 
 #include "aether/memory_buffer.h"
-#include "aether/transport/data_buffer.h"
+#include "aether/types/data_buffer.h"
 
 namespace ae {
